Put SymbolMatcher/NameResolver to separate directory
