+++ ModelBase/src/nodes/Node.h
@@ -28,7 +28,7 @@
 
 #include "../modelbase_api.h"
 #include "../persistence/PersistentStore.h"
-#include "../SymbolMatcher.h"
+#include "../util/SymbolMatcher.h"
 #include "../model/AllTreeManagers.h"
 
 #include "Core/src/reflect/Reflect.h"
